Added Late Check
Added a Late assignment check that checks to see if the assignment is late. This info is put into the code as variable 'late' and will turn date read if the assignment is late. Also added basic current time to navbar but is offcentered and needs to be fixed. The time is only UTC so we will need to add change to local time