Various Bug Fixes and Featured
Added the score on the teachers grading page.

Fixed the html showing in the instructions of submissions

The past few classes we were trying to properly deploy our project. We concluded that formally deploying our project is too much work so we decided that focus on bug fixes additional features instead of a unimportant part of our website. This is why there are not many entries for the last few classes.